% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_diversity.R
\docType{data}
\name{language_diversity}
\alias{language_diversity}
\title{Language diversity data from Nettle (1998).}
\format{
A data frame with 444 rows and 4 variables:
\describe{
  \item{Continent}{continent lables}
  \item{Country}{country labels}
  \item{Measurement}{global pattern measures}
  \item{Value}{values of measurements}
}
}
\source{
\url{https://github.com/jvcasillas/untidydata}
}
\usage{
language_diversity
}
\description{
A long-form dataset containing language diversity observations from Nettle 
(1998). The data attempt to explain global patterns of language diversity 
and are most useful in wide format. 
 
Nettle, D. (1998). Explaining Global Patterns of Language Diversity.
Journal of Anthropological Archaeology, 17, 354–374.
}
\keyword{datasets}
