% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_post.R
\docType{data}
\name{pre_post}
\alias{pre_post}
\title{Pre/post test data from two populations.}
\format{
A data frame with 16 rows and 4 variables:
\describe{
  \item{id}{subject id labels, language label}
  \item{spec}{two grouping variables separated by underscores}
  \item{test1}{scores from pre-test}
  \item{test2}{scores from post-test}
}
}
\source{
\url{https://github.com/jvcasillas/untidydata}
}
\usage{
pre_post
}
\description{
A simulated dataset containing untidy test scores in percents
from two groups. This data was simulated in R for 
teaching purposes.
}
\keyword{datasets}
