% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirantization.R
\docType{data}
\name{spirantization}
\alias{spirantization}
\title{Spanish consonant and vowel intensity measurements.}
\format{
A data frame with 16 rows and 4 variables:
\describe{
  \item{id}{subject id and proficiency label}
  \item{position}{word initial or intervocalic position}
  \item{cIntensity}{intensity in dB from obstruent}
  \item{vIntensity}{intensity in dB from following vowel}
}
}
\source{
\url{https://github.com/jvcasillas/untidydata}
}
\usage{
spirantization
}
\description{
A simulated dataset containing intensity measurements from a word 
initial or intervocalic obstruent-vowel sequence for two learner 
groups and native Spanish speakers.
}
\keyword{datasets}
