% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vot.R
\docType{data}
\name{vot}
\alias{vot}
\title{Voice onset time data from three populations.}
\format{A data frame with 720 rows and 5 variables:
\describe{
  \item{participant}{subject id labels}
  \item{language}{language spoken, spanish/english}
  \item{item}{target word}
  \item{repitition}{iteration of item}
  \item{vot}{vot values, in ms}
}}
\source{
\url{http://www.jvcasillas.com/}
}
\usage{
vot
}
\description{
A dataset containing raw voice onset time (VOT) data in ms 
from three groups. This data was simulated in R based of 
the data set used in Casillas, Diaz, & Simonet (2015).
 
Casillas, J. V., Diaz, Y. and Simonet, M. (2015). "Acoustics of 
Spanish and English coronal stops". In ICPhS 18. Ed. by T. S. C. 
for ICPhS 2015. Proceedings of the 18th International Congress 
of Phonetic Sciences. University of Glasgow: Glasgow. 
http://www.icphs2015.info, ISSN: 241-0669, pp. 1-5.
}
\keyword{datasets}
